Decompose aggregate_formatter into 6 focused files
Extract structure_formatters, behavior_formatters, rule_formatters,
lifecycle_formatter as includable modules. Phase 3 of HEC-493.
